Update the type of nummonbssid field in the access_points table

+++ database/schema/mysql-schema.sql
@@ -44,7 +44,7 @@ CREATE TABLE `access_points` (
   `numasoclients` smallint(6) NOT NULL DEFAULT 0,
   `nummonclients` smallint(6) NOT NULL DEFAULT 0,
   `numactbssid` tinyint(4) NOT NULL DEFAULT 0,
-  `nummonbssid` tinyint(4) NOT NULL DEFAULT 0,
+  `nummonbssid` int(11) NOT NULL DEFAULT 0,
   `interference` tinyint(3) unsigned NOT NULL,
   PRIMARY KEY (`accesspoint_id`),
   KEY `name` (`name`,`radio_number`),
